+++ src/components/dropdown/dropdown.spec.js
@@ -95,6 +95,14 @@ describe('dropdown', async () => {
   })
   */
 
+  it('should have a toggle with the given toggle tag', async () => {
+    const { app: { $refs } } = window
+    const { dd_10 } = $refs // eslint-disable-line camelcase
+
+    const toggle = dd_10.$el.querySelector('.dropdown-toggle')
+    expect(toggle).toBeElement('div')
+  })
+
   it('dd-item should render as link by default', async () => {
     const { app: { $refs } } = window
     const { dd_6 } = $refs // eslint-disable-line camelcase